ChatBot/requirements.txt
pypdf
docx2txt
openai
python-dotenv
gpt_index
gradio